+++ src/main/kotlin/mdnet/base/Constants.kt
@@ -1,7 +1,6 @@
 package mdnet.base
 
-import com.google.gson.Gson
-import com.google.gson.GsonBuilder
+import com.fasterxml.jackson.module.kotlin.jacksonObjectMapper
 import java.time.Duration
 
 object Constants {
@@ -10,5 +9,5 @@ object Constants {
     const val WEBUI_VERSION = "0.1.1"
     val MAX_AGE_CACHE: Duration = Duration.ofDays(14)
     @JvmField
-    val GSON: Gson = GsonBuilder().setPrettyPrinting().create()
+    val JACKSON = jacksonObjectMapper()
 }